+++ CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## [0.19.1](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.19.0...v0.19.1) (2025-10-19)
+
+
+### Bug Fixes
+
+* add aria-label to Icon ([9d484e4](https://github.com/tuyuritio/astro-theme-thought-lite/commit/9d484e4a50c6cb44f3b83e5d6bb0c6e4b8cfc26f))
+* **style:** update dark theme background ([46b7d42](https://github.com/tuyuritio/astro-theme-thought-lite/commit/46b7d4253e0952e8b039daa81c0433adc386d328))
+* **style:** update jotting border color ([bffc75d](https://github.com/tuyuritio/astro-theme-thought-lite/commit/bffc75d1a8809f914c8caa87dfb6c373183ae429))
+
+## [0.19.0](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.18.0...v0.19.0) (2025-10-17)
+
+
+### Features
+
+* add interactive CLI script for creating new content ([dadd562](https://github.com/tuyuritio/astro-theme-thought-lite/commit/dadd562ca76001e898323a39de03d1f76f3a74cf))
+
+
+### Bug Fixes
+
+* **jotting:** adjust border thickness ([25ffb7e](https://github.com/tuyuritio/astro-theme-thought-lite/commit/25ffb7edd21a80349d0791a7ff68da978087decb))
+
+## [0.18.0](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.17.3...v1.0.0) (2025-10-13)
+
+
+### ⚠ BREAKING CHANGES
+
+* **note:** `contents` is no longer supported in frontmatter, use `toc` instead.
+
+### Code Refactoring
+
+* **note:** rename frontmatter `contents` field to `toc` ([6a3673a](https://github.com/tuyuritio/astro-theme-thought-lite/commit/6a3673a2202dab4f7f545516806b0a4f734a1f5a))
+
+## [0.17.3](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.17.2...v0.17.3) (2025-10-08)
+
+
+### Bug Fixes
+
+* **style:** revert font style ([94445e9](https://github.com/tuyuritio/astro-theme-thought-lite/commit/94445e9b6c075151366d2fd4d50aa86006697c31))
+
+## [0.17.2](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.17.1...v0.17.2) (2025-10-08)
+
+
+### Bug Fixes
+
+* **style:** adjust font style ([a9fee39](https://github.com/tuyuritio/astro-theme-thought-lite/commit/a9fee39367d4827976a2e70ea815f7ea4e60c888))
+* **style:** wrap headings with link instead of appending ([08ab988](https://github.com/tuyuritio/astro-theme-thought-lite/commit/08ab988f3890423b9d5f22b879b12be34246e42a))
+
+## [0.17.1](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.17.0...v0.17.1) (2025-10-06)
+
+
+### Bug Fixes
+
+* add missing paginator styles ([249d39b](https://github.com/tuyuritio/astro-theme-thought-lite/commit/249d39b0fadedd7f56fcd9ddb0cef191cc201b27))
+* set initial values for list rendering ([6300997](https://github.com/tuyuritio/astro-theme-thought-lite/commit/630099735ec515347916d050b6424086dafd738c))
+* **style:** make hover effect responsively ([8f334d0](https://github.com/tuyuritio/astro-theme-thought-lite/commit/8f334d03bd19c1473824ec22b66ac61085633387))
+
 ## [0.17.0](https://github.com/tuyuritio/astro-theme-thought-lite/compare/v0.16.0...v0.17.0) (2025-10-05)